details render errors
