+++ Finjector.Web/ClientApp/src/components/Shared/Layout/PageTitle.tsx
@@ -1,10 +1,11 @@
 import React from "react";
 interface PageTitleProps extends React.HTMLAttributes<HTMLDivElement> {
   title?: string;
+  isRow?: boolean;
 }
 
 const PageTitle: React.FC<PageTitleProps> = (
-  { title, children, className },
+  { title, children, className, isRow = false },
   props
 ) => {
   if (!children && !title) {
@@ -12,7 +13,7 @@ const PageTitle: React.FC<PageTitleProps> = (
   }
   let divClassName = `page-title mb-3 pb-2 ${className ?? ""}`;
 
-  if (!title && Array.isArray(children) && children.length > 1) {
+  if ((!title && Array.isArray(children) && children.length > 1) || isRow) {
     // this will not be applied if children is a single element (such as <h1>, <div>, <>) or if title is set
     // but if we have multiple elements, put them in a row and apply the correct styles
     divClassName += " row justify-content-between align-items-center";
