--- Finjector.Web/ClientApp/src/components/Shared/LoadingAndErrors/ChartNotFound.tsx
@@ -1,8 +1,0 @@
-export const ChartNotFound = () => {
-  return (
-    <div className="p-2">
-      <h1>Chart Not Found</h1>
-      <p>Please check that you have a valid chart ID and try again.</p>
-    </div>
-  );
-};
