+++ Finjector.Web/ClientApp/src/pages/Details.tsx
@@ -1,39 +1,20 @@
-import React from "react";
-import { useParams, useNavigate } from "react-router-dom";
-import FinLoader from "../components/Shared/LoadingAndErrors/FinLoader";
-
-import { AeDetails, ChartStringEditModel, ChartType } from "../types";
+import { useParams } from "react-router-dom";
+import { AeDetails, ChartType } from "../types";
 import { useGetChartDetails } from "../queries/storedChartQueries";
 import { ChartDebugInfo } from "../components/Shared/LoadingAndErrors/ChartDebugInfo";
-import { ChartLoadingError } from "../components/Shared/LoadingAndErrors/ChartLoadingError";
-import { Alert } from "reactstrap";
-import { renderNameAndEmail } from "../util/util";
-import { ChartNotFound } from "../components/Shared/LoadingAndErrors/ChartNotFound";
-import DetailsChartString from "../components/Details/DetailsChartString";
-import PpmDetailsPage from "../components/Details/PpmDetails";
-import FinButton from "../components/Shared/FinButton";
-import SharePopup from "../components/Details/SharePopup";
-import CopyToClipboardHover from "../components/Shared/CopyToClipboardHover";
-import { DetailsRow } from "../components/Details/DetailsRow";
-
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { faBolt, faPencil } from "@fortawesome/free-solid-svg-icons";
-import usePopupStatus from "../util/customHooks";
+import DetailsTitle from "../components/Details/DetailsTitle";
+import DetailsBody from "../components/Details/DetailsBody";
 import PageTitle from "../components/Shared/Layout/PageTitle";
-import PageBody from "../components/Shared/Layout/PageBody";
+import DetailsPageBody from "../components/Shared/Layout/DetailsPageBody";
 
 const Details = () => {
-  const { chartId, teamId, folderId, chartSegmentString } = useParams();
+  const { chartId, chartSegmentString } = useParams();
   const chartDetailsQuery = useGetChartDetails(
     chartSegmentString || "",
     chartId
   );
 
   const aeDetails: AeDetails | undefined = chartDetailsQuery.data?.aeDetails;
-  const chartStringDetails: ChartStringEditModel | undefined =
-    chartDetailsQuery.data?.chartStringDetails;
-  const isInPopup = usePopupStatus();
-  const navigate = useNavigate();
 
   const invalid =
     (chartDetailsQuery.isLoading && chartDetailsQuery.isFetching) || // if we're doing first fetch
@@ -41,174 +22,28 @@ const Details = () => {
     !aeDetails?.chartString || // if we have no data
     aeDetails.chartType === ChartType.INVALID; // if we have invalid data
 
-  const getEditLinkUrl = () => {
-    if (chartId) {
-      return `/teams/${teamId}/folders/${folderId}/entry/${chartId}/${aeDetails?.chartString}`;
-    } else {
-      return `/entry/${aeDetails?.chartString}`;
-    }
-  };
-
-  const renderLoadingOrError = () => {
-    if (chartDetailsQuery.isLoading && chartDetailsQuery.isFetching) {
-      return (
-        <div className={`chartstring-details is-none`}>
-          <FinLoader />
-        </div>
-      );
-    }
-    if (chartDetailsQuery.isError) {
-      return (
-        <div className={`chartstring-details is-none`}>
-          <ChartLoadingError />
-        </div>
-      );
-    }
-    if (
-      !chartSegmentString ||
-      !aeDetails?.chartString ||
-      aeDetails.chartType === ChartType.INVALID
-    ) {
-      return (
-        <div className={`chartstring-details is-none`}>
-          <ChartNotFound />
-        </div>
-      );
-    }
-  };
-  const isPpmOrGlClassName =
-    aeDetails?.chartType === ChartType.PPM ? "is-ppm" : "is-gl";
-  const use = () => {
-    if (chartId) {
-      navigate(
-        `/teams/${teamId}/folders/${folderId}/selected/${chartId}/${aeDetails?.chartString}`
-      );
-    } else {
-      navigate(`/selected/${aeDetails?.chartString}`);
-    }
-  };
-
   return (
     <div className="main">
-      <PageTitle>
-        <div className="col-12 col-md-7">
-          {chartStringDetails && (
-            <h4>
-              {chartStringDetails.teamName} {"/ "}
-              {chartStringDetails.folder?.name}
-            </h4>
-          )}
-          {chartStringDetails?.name === undefined ? (
-            <h1>Chart String Details</h1>
-          ) : (
-            <CopyToClipboardHover
-              value={chartStringDetails.name}
-              id="ChartName"
-            >
-              <h1>{chartStringDetails.name}</h1>
-            </CopyToClipboardHover>
-          )}
-        </div>
-        {!invalid && (
-          <div className="col-12 col-md-5 text-end">
-            {isInPopup && (
-              <FinButton onClick={use}>
-                <FontAwesomeIcon icon={faBolt} />
-                Use
-              </FinButton>
-            )}
-            <FinButton to={getEditLinkUrl()}>
-              <FontAwesomeIcon icon={faPencil} />
-              Edit Chart String
-            </FinButton>
-            <SharePopup chartString={aeDetails.chartString} />
-          </div>
-        )}
+      <PageTitle isRow={true}>
+        <DetailsTitle
+          aeDetails={aeDetails}
+          chartStringDetails={chartDetailsQuery.data?.chartStringDetails}
+          invalid={invalid}
+          isLoading={chartDetailsQuery.isLoading}
+          isFetching={chartDetailsQuery.isFetching}
+          isError={chartDetailsQuery.isError}
+        />
       </PageTitle>
-      <PageBody>
-        {!!aeDetails && (
-          <div>
-            {aeDetails.errors.length > 0 &&
-              aeDetails.errors.map((error, i) => {
-                return (
-                  <Alert color="danger" key={i}>
-                    Error: {error}
-                  </Alert>
-                );
-              })}
-            {!!aeDetails &&
-              aeDetails.hasWarnings &&
-              aeDetails.warnings.length > 0 &&
-              aeDetails.warnings.map((warning, i) => {
-                return (
-                  <Alert color="warning" key={i}>
-                    Warning: {warning}
-                  </Alert>
-                );
-              })}
-          </div>
-        )}
-        {invalid ? (
-          renderLoadingOrError()
-        ) : (
-          <div className={`chartstring-details ${isPpmOrGlClassName}`}>
-            <DetailsChartString
-              chartType={aeDetails.chartType}
-              chartString={aeDetails.chartString}
-              isValid={aeDetails.isValid}
-              hasWarnings={aeDetails.hasWarnings}
-            />
-            <div className="chartstring-details-info unique-bg">
-              {aeDetails.segmentDetails.map((segment, i) => {
-                return (
-                  <DetailsRow
-                    headerColText={segment.entity}
-                    key={i}
-                    column2={
-                      <span className="fw-bold primary-font me-4">
-                        <CopyToClipboardHover
-                          value={segment.code ?? ""}
-                          id={`segment-code-${i}`}
-                        >
-                          {segment.code ?? ""}{" "}
-                        </CopyToClipboardHover>
-                      </span>
-                    }
-                    column3={
-                      <CopyToClipboardHover
-                        value={segment.name ?? ""}
-                        id={`segment-name-${i}`}
-                      >
-                        {segment?.name ?? ""}
-                      </CopyToClipboardHover>
-                    }
-                  />
-                );
-              })}
-            </div>
-            <div className="chartstring-details-info">
-              {aeDetails.chartType === ChartType.PPM && (
-                <PpmDetailsPage details={aeDetails.ppmDetails} />
-              )}
-              <DetailsRow
-                headerColText="GL Financial Department SCM Approver(s)"
-                column2={aeDetails.approvers.map((approver, i) => {
-                  return (
-                    <div key={i}>
-                      <CopyToClipboardHover
-                        value={approver.email ?? ""}
-                        id={`approver-${i}`}
-                      >
-                        {renderNameAndEmail(approver.name, approver.email)}
-                      </CopyToClipboardHover>
-                    </div>
-                  );
-                })}
-              />
-            </div>
-          </div>
-        )}
-      </PageBody>
+      <DetailsPageBody>
+        <DetailsBody
+          aeDetails={aeDetails}
+          chartSegmentString={chartSegmentString}
+          invalid={invalid}
+          isLoading={chartDetailsQuery.isLoading}
+          isFetching={chartDetailsQuery.isFetching}
+          isError={chartDetailsQuery.isError}
+        />
+      </DetailsPageBody>
       <ChartDebugInfo chartDetails={chartDetailsQuery.data} />
     </div>
   );
